+++ .gitignore
@@ -44,3 +44,11 @@ next-env.d.ts
 .eslintcache
 
 *.zip*
+
+
+# Playwright
+node_modules/
+/test-results/
+/playwright-report/
+/blob-report/
+/playwright/.cache/
