+++ src/shared/lib/api/fetcher/fetcher.ts
@@ -1,4 +1,4 @@
-import { handleAuthError } from '@shared/config/tanstack-query/mutation-defaults'
+import { handleAuthError } from '@shared/config/tanstack-query/utils'
 import { EndpointType } from '@shared/constants/endpoint/endpoint'
 
 import { refreshAccessToken } from './reissue'
@@ -53,13 +53,11 @@ export async function fetchRequest<T>(
       } catch (refreshError) {
         const error: ErrorResponse = {
           status: 401,
-          message: '세션이 만료되었습니다. 다시 로그인해주세요.',
+          message: '세션이 만료되었습니다. 다시 로그인해주세요',
           data: null,
         }
         /** 미인증 유저에게는 토스트로 알림 */
-        if (typeof window !== 'undefined') {
-          handleAuthError(error)
-        }
+        handleAuthError(error)
         throw error
       }
     }
